+++ NEWS.md
@@ -8,7 +8,7 @@ Version 1.1.1 - beta
 - NEW: Python 3.12 binary wheel deployment on PyPI
 - NEW: page alignment support in the `drawing` add-on for these backends: 
   `SVGBackend`, `PyMuPdfBackend` and the `PlotterBackend`
-- NEW: support for decoding of MIF encoded text `\M+cxxxx` in the `recover` module
+- NEW: support for decoding of MIF encoded text `\M+cxxxx` by the `recover` module
 - INFO: `numpy` v1.25 has stopped providing Python 3.8 binary wheels on PyPI
 - BUGFIX: [#929](https://github.com/mozman/ezdxf/issues/929)
   handling of the minimum hatch line distance